% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogPackFunc.R
\name{dogPack_demo_setup}
\alias{dogPack_demo_setup}
\title{Perform Setup}
\usage{
dogPack_demo_setup(where = NULL)
}
\arguments{
\item{where}{Character length one. Holding the path to the folder where the 
local settings.R file should be stored.}
}
\value{
No return value, demo function called for its side effects, i.e. to
perform the setup.
}
\description{
Set up the local system for the use of the dynamic
settings file.
}
