% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogPackFunc.R
\name{dogPack_demo_tellFavouriteColor}
\alias{dogPack_demo_tellFavouriteColor}
\title{Tell favourite Color}
\usage{
dogPack_demo_tellFavouriteColor()
}
\value{
Character length one, the favourite color as defined in the 
settings.R file.
}
\description{
Tells the favourite color as defined in the settings.R file.
}
