% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogPackFunc.R
\name{dogPack_demo_updateSettings}
\alias{dogPack_demo_updateSettings}
\title{Demo Update Settings}
\usage{
dogPack_demo_updateSettings(silent = FALSE)
}
\arguments{
\item{silent}{Logical. If a confirmation should be printed. Defaults to 'FALSE'.}
}
\value{
A (invisible) list containing the key=value pairs from the settings file.
}
\description{
Update Settings
}
