% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogPackFunc.R
\name{dogPack_test_targetPackageParams}
\alias{dogPack_test_targetPackageParams}
\title{Test Target Package Parameters}
\usage{
dogPack_test_targetPackageParams()
}
\value{
Prints target package name, environment name, name of list-object,
name of folder where the local 'dogPack_settings.R' resides etc, 
and gives it back as (invisible) list.
}
\description{
Simple test if setup is ok.
}
