% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogPackFunc.r
\name{dogPackTest}
\alias{dogPackTest}
\title{dogPack TEst}
\usage{
dogPackTest()
}
\value{
Prints target package name, environment name, name of list-object,
name of folder where the local 'dogPack_settings.R' resides etc.
}
\description{
Simple test if setup is ok.
}
