% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogPackFunc.r
\name{dogPack_demo_autoUpS}
\alias{dogPack_demo_autoUpS}
\title{Test Autoupdate}
\usage{
dogPack_demo_autoUpS()
}
\description{
AutoUpdate Settings. Is also reading the value of the
key 'favouriteColor' from the 'dogPack_settings.R' file.
}
