% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogPackFunc.R
\name{dogPack_demo_autoUpS}
\alias{dogPack_demo_autoUpS}
\title{Demo Autoupdate}
\usage{
dogPack_demo_autoUpS(tellMe = TRUE, txt = "My favourite color is ")
}
\arguments{
\item{tellMe}{Logical. If the favourite color should be printed.}

\item{txt}{Character}
}
\value{
Is called for its side effects, i.e. to automatically update 
resp. (re-)source the settings file. If 'tellMe' is left at the default 
'TRUE', the value of the key 'favouriteColor' in the settings file is 
printed. Returns (invisible) NULL.
}
\description{
AutoUpdate Settings. Is also reading the value of the
key 'favouriteColor' from the 'dogPack_settings.R' file.
}
